requirements.txt
discord.py>=2.3.2
yt-dlp>=2023.12.30
PyNaCl>=1.5.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0